v0.0.3, support larger inputs
According to report:

> I have just been trying to use your gzip node to unzip downloaded XML
files, which can be a few megabytes big. Even with fairly small files I
was getting a "stack exhausted" error on the decode.

The error came from the line:

var s = String.fromCharCode.apply(null, out);

It appears there is a maximum limit on the size of the input array of
about 250k.

Instead, TextDecoder works well with larger inputs.

Signed-off-by: Colin Smale <colin.smale@xs4all.nl>